+++ src/main/java/io/vntr/spaja/SpajaMiddleware.java
@@ -167,12 +167,12 @@ public class SpajaMiddleware implements IMiddleware, IMiddlewareAnalyzer {
 
     @Override
     public Integer getNumberOfPartitions() {
-        return (int) manager.getAllPartitionIds().size();
+        return manager.getAllPartitionIds().size();
     }
 
     @Override
     public Integer getNumberOfUsers() {
-        return (int) manager.getNumUsers();
+        return manager.getNumUsers();
     }
 
     @Override
